requirements.txt
fastapi
uvicorn[standard]
deepgram-sdk==2.12.0
elevenlabs==1.13.5
twilio
openai
python-dotenv
